+++ src/lib/components/layouts/Footer.svelte
@@ -1,0 +1,4 @@
+<footer class="px-16 py-6 w-full flex justify-between text-[#5f5f5f] font-eurostile">
+  <p>A developer</p>
+  <p>© Kz Creation</p>
+</footer>