+++ src/lib/components/three/Index.svelte
@@ -3,7 +3,7 @@
   import Scene from './Scene.svelte';
 </script>
 
-<div class="fixed inset-0 -z-5 pointer-events-none">
+<div class="fixed inset-0 -z-10 pointer-events-none">
   <Canvas>
     <Scene />
   </Canvas>