+++ src/routes/+page.svelte
@@ -2,7 +2,9 @@
   import Menu from '$lib/components/layouts/Menu.svelte';
 </script>
 
-<main class="px-[5vw] py-[10vh] h-screen flex justify-between">
-  <h1 class="text-3xl font-futura_pt">Welcome to Kz Creation portfolio.</h1>
-  <Menu />
+<main class="flex justify-between px-[5vw] py-[10vh] h-screen">
+  <div class="flex flex-col items-start w-full min-h-screen text-white gap-8">
+    <h1 class="text-3xl font-futura_pt tracking-wider">Welcome to Kz Creation portfolio.</h1>
+    <Menu />
+  </div>
 </main>